Restore retired docs redirects
Handle retired memory docs URLs in the docs Worker before static asset lookup and validate that they redirect without touching ASSETS. Share the redirect map with sitemap filtering so runtime and SEO behavior stay in sync.
